fix: revert helper signature to usize to match constant and len() type

+++ creator-keys/tests/events.rs
@@ -101,7 +101,7 @@ fn assert_event_topic_matches(env: &Env, event: &(Address, Vec<Val>, Val), expec
     );
 }
 
-fn assert_event_field_count(event: &(Address, Vec<Val>, Val), env: &Env, expected: u32) {
+fn assert_event_field_count(event: &(Address, Vec<Val>, Val), env: &Env, expected: usize) {
     let data_vec: Vec<Val> = event.2.clone().into_val(env);
     assert_eq!(
         data_vec.len(),
